+++ src/main/kotlin/com/krillsson/sysapi/core/domain/docker/ComposeLabels.kt
@@ -1,0 +1,5 @@
+package com.krillsson.sysapi.core.domain.docker
+
+object ComposeLabels {
+    const val PROJECT = "com.docker.compose.project"
+}